fix: 두 개 이상의 로또가 일등 당첨될 경우 상금이 Integer 범위를 넘어가는 문제 수정

+++ src/main/java/lotto/domain/LottoResult.java
@@ -17,14 +17,14 @@ public class LottoResult {
     }
 
     public double calculateEarningRate() {
-        double totalEarning = 0;
+        long totalEarning = 0;
         int totalCount = 0;
 
         for (WinningPrize winningPrize : WinningPrize.values()) {
             totalCount += winningInformation.get(winningPrize);
             totalEarning += (winningPrize.getPrize() * winningInformation.get(winningPrize));
         }
-        return (totalEarning / (totalCount * 1000)) * 100;
+        return ((double) totalEarning / (totalCount * 1000)) * 100;
     }
 
     public Map<WinningPrize, Integer> getWinningInformation() {
